task: Move callback handling logic to task.Runner
