+++ clients/catalyst.go
@@ -1,11 +1,9 @@
-package catalyst
+package clients
 
 import (
 	"bytes"
 	"context"
 	"encoding/json"
-
-	"github.com/livepeer/task-runner/clients"
 )
 
 type UploadVODRequest struct {
@@ -21,13 +19,27 @@ type OutputLocation struct {
 	PinataAccessKey string `json:"pinata_access_key"`
 }
 
-type Client interface {
+type CatalystCallback struct {
+	Status          string       `json:"status"`
+	CompletionRatio float64      `json:"completion_ratio"`
+	Error           string       `json:"error"`
+	Retriable       bool         `json:"retriable"`
+	Outputs         []OutputInfo `json:"outputs"`
+}
+
+type OutputInfo struct {
+	Type     string            `json:"type"`
+	Manifest string            `json:"manifest"`
+	Videos   map[string]string `json:"videos"`
+}
+
+type Catalyst interface {
 	UploadVOD(ctx context.Context, upload UploadVODRequest) error
 }
 
-func NewClient(url, authSecret string) Client {
-	return &client{
-		BaseClient: clients.BaseClient{
+func NewCatalyst(url, authSecret string) Catalyst {
+	return &catalyst{
+		BaseClient: BaseClient{
 			BaseUrl: url,
 			BaseHeaders: map[string]string{
 				"Authorization": "Bearer " + authSecret,
@@ -36,16 +48,16 @@ func NewClient(url, authSecret string) Client {
 	}
 }
 
-type client struct {
-	clients.BaseClient
+type catalyst struct {
+	BaseClient
 }
 
-func (c *client) UploadVOD(ctx context.Context, upload UploadVODRequest) error {
+func (c *catalyst) UploadVOD(ctx context.Context, upload UploadVODRequest) error {
 	body, err := json.Marshal(upload)
 	if err != nil {
 		return err
 	}
-	return c.DoRequest(ctx, clients.Request{
+	return c.DoRequest(ctx, Request{
 		Method:      "POST",
 		URL:         "/api/vod",
 		Body:        bytes.NewReader(body),
